fix: use Int64 for session timestamps in iCloud snapshot
The `toolDetails` session records carry epoch-millisecond timestamps in
`firstTs`/`lastTs`, but both the macOS writer and the iOS/watchOS reader
declared them as `Int`. On watchOS (arm64_32) `Int` is 32-bit, so decoding
those ~1.7e12 values overflows and the whole DashboardSnapshot decode fails,
leaving watchOS showing all zeros.

+++ Sources/Store/DashboardCache.swift
@@ -159,8 +159,8 @@ struct ToolSessionItem: Codable {
     var sessionId: String?
     var title: String?
     var repo: String?
-    var firstTs: Int
-    var lastTs: Int
+    var firstTs: Int64
+    var lastTs: Int64
     var cost: Double
     var windowTokens: Int?
     var lastInput: Int
